Undeprecate preview traits
